Make seperated file about removing when multi section selected
